Huseyin validator

+++ src/main/java/com/garanti/TeknikServis/controller/BookingController.java
@@ -1,20 +1,24 @@
 package com.garanti.TeknikServis.controller;
 import com.garanti.TeknikServis.model.Booking;
 import com.garanti.TeknikServis.model.BookingDTO;
+import com.garanti.TeknikServis.model.Users;
 import com.garanti.TeknikServis.repo.BookingRepo;
+import com.garanti.TeknikServis.response.RestResponse;
 import com.garanti.TeknikServis.service.BookingService;
+import com.garanti.TeknikServis.service.UserService;
+import com.garanti.TeknikServis.validation.BookingValidator;
+import org.springframework.context.MessageSource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Locale;
 
 @RestController
 @RequestMapping(path = "appointment")
@@ -22,16 +26,24 @@ public class BookingController {
 
     private BookingService appointmentService;
     private BookingRepo appointmentRepo;
+    private UserService userService;
 
-    public BookingController(BookingService appointmentService, BookingRepo appointmentRepo) {
+    private MessageSource messageSource;
+
+    private BookingValidator bookingValidator;
+
+    public BookingController(BookingService appointmentService, BookingRepo appointmentRepo, UserService userService, MessageSource messageSource, BookingValidator bookingValidator) {
         this.appointmentService=appointmentService;
         this.appointmentRepo=appointmentRepo;
+        this.userService = userService;
+        this.messageSource= messageSource;
+        this.bookingValidator = bookingValidator;
     }
 
-    @Secured("ROLE_USER")
+    @PreAuthorize(value = "isAuthenticated()")
     @GetMapping(path = "getById", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<BookingDTO> getByIdQueryParam(@RequestParam(value = "id", required = true) Integer id) {
-        // localhost:9090/FirstSpringWeb/ogretmen/getById?id=1
+        // http://localhost:9090/appointment/getById?id=1
         BookingDTO res = appointmentService.getById(id);
         if (res != null) {
             return ResponseEntity.ok(res);
@@ -42,30 +54,75 @@ public class BookingController {
 
 
     @PostMapping(path = "save", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<String> save(@RequestBody Booking booking)
+    public ResponseEntity save(@Validated @RequestBody Booking booking, BindingResult result, @RequestHeader(name="Accept-Language", required = false) Locale locale)
     {
-        //appointmentService.deneme(booking);
+        // {"note":"Formatlamaistiyorum", "user_ID":2, "service_ID":3}
+        //localhost:9090/appointment/save
+
+        bookingValidator.validate(booking, result);
+
+        if (result.hasErrors()) {
+            return ResponseEntity.badRequest().body(result.getAllErrors());
+        }
 
         if (appointmentService.save(booking))
         {
-
-            return ResponseEntity.status(HttpStatus.CREATED).body("Kaydınız oluşturulmuştur.Kayıt numaranız: "+appointmentRepo.getNextId()+"dır.Lütfen kaybetmeyiniz.!");
+            return ResponseEntity.status(HttpStatus.CREATED).body(messageSource.getMessage("booking.save.success",null, locale)+ appointmentRepo.getNextId());
         }
         else
         {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Kaydınızda bir sorun oluştu. Lütfen yeniden kayıt oluşturunuz.");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(messageSource.getMessage("booking.save.fail",null, locale));
         }
     }
 
-    @Secured("ROLE_USER")
+    @PreAuthorize("isAuthenticated()")
     @DeleteMapping(path = "deleteById/{id}")
-    public ResponseEntity<String> deleteById(@PathVariable(value = "id") Integer id) {
-        // localhost:9090/FirstSpringWeb/ogretmen/deleteById/1
-        if (appointmentService.deleteById(id)) {
-            return ResponseEntity.ok("Başarı ile silindi");
-        } else {
-            return ResponseEntity.internalServerError().body("Başarı ile silinemedi");
+    public ResponseEntity<String> deleteById(Authentication auth, @PathVariable(value = "id") Integer id,@RequestHeader(name = "Accept-Language", required = false) Locale locale) {
+        // http://localhost:9090/appointment/deleteById/12
+        if (auth.isAuthenticated())
+        {
+            System.err.println("-------> yes auth");
         }
+        else
+        {
+            // buraya düşeceğini sanmıyoruz
+            System.err.println("-------> not auth");
+        }
+        String username = auth.getPrincipal().toString();
+        Users usr = userService.getUserByUsername(username);
+        if (!usr.getUSER_ID().equals(appointmentRepo.getBookingById(id).getUSER_ID()))
+        {
+            return ResponseEntity.badRequest().body(messageSource.getMessage("booking.wrong.user",null,locale));
+        }
+
+        if (appointmentService.deleteById(id)) {
+            return ResponseEntity.ok(messageSource.getMessage("booking.delete.success",null,locale));
+        } else {
+            return ResponseEntity.internalServerError().body(messageSource.getMessage("booking.delete.fail", null, locale));
+        }
+    }
+
+    @GetMapping("getAppointmentsDatesInOrder")
+    @Secured(value = "ROLE_ADMIN")
+    public ResponseEntity getAppointmentsDatesInOrder(@RequestParam(name = "sortType") String type,@RequestHeader(name = "Accept-Language", required = false) Locale locale){
+        //http://localhost:9090/appointment/getAppointmentsDatesInOrder?sortType=asc
+        return ResponseEntity.ok(RestResponse.of(appointmentService.getAppointmentsDatesInOrder(type, locale)));
+    }
+
+    //http://localhost:9090/appointment/getAllAppointmentLikeUsername?username=A
+    @GetMapping("getAllAppointmentLikeUsername")
+    @Secured(value = "ROLE_ADMIN")
+    public ResponseEntity getAllAppointmentLikeUsername (@RequestParam(name = "username") String username, @RequestHeader(name = "Accept-Language", required = false)Locale locale){
+        return ResponseEntity.ok(RestResponse.of(appointmentService.getAllAppointmentLikeUsername(username, locale)));
+    }
+
+
+
+    @PatchMapping("appointmentIsComplete")
+    @Secured(value = "ROLE_ADMIN")
+    public ResponseEntity<?> appointmentIsComplete(@RequestParam(name = "id") int id, @RequestParam (name = "is_done") boolean is_done,@RequestHeader(name = "Accept-Language", required = false) Locale locale){
+        //http://localhost:9090/appointment/appointmentIsComplete?id=1&is_done=true
+        return ResponseEntity.ok(RestResponse.of(appointmentService.appointmentIsComplete(id,is_done, locale)));
     }
